Writing Hello World within a Command Line
Learned how to WriteLine a line of text with in the command line. Console.Read() added to allow the .exe to be viewable outside of Visual Studio. This is written in .Net5. I will upload a .Net6 at a later point to compare and contrast the differences between the two.